requirements.txt
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
bcrypt==4.2.1
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
croniter==6.0.0
cryptography==44.0.1
debtcollector==3.0.0
decorator==5.1.1
dnspython==2.7.0
dogpile.cache==1.3.4
eval_type_backport==0.2.2
eventlet==0.39.0
fasteners==0.19
futurist==3.0.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
iso8601==2.1.0
Jinja2==3.1.5
jsonpath-python==1.0.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keystoneauth1==5.9.1
keystonemiddleware==10.8.0
kombu==5.4.2
legacy-cgi==2.6.2
logutils==0.3.5
Mako==1.3.9
MarkupSafe==3.0.2
mistral-lib==3.3.0
mistralai==1.5.0
msgpack==1.1.0
mypy-extensions==1.0.0
netaddr==1.3.0
networkx==3.4.2
os-service-types==1.7.0
oslo.cache==3.9.0
oslo.concurrency==7.0.0
oslo.config==9.7.0
oslo.context==5.7.0
oslo.db==17.1.0
oslo.i18n==6.5.0
oslo.log==7.0.0
oslo.messaging==16.0.0
oslo.metrics==0.10.1
oslo.middleware==6.3.0
oslo.policy==4.5.0
oslo.serialization==5.6.0
oslo.service==4.0.0
oslo.utils==8.0.0
osprofiler==4.2.0
packaging==24.2
paramiko==3.5.1
Paste==3.10.1
PasteDeploy==3.1.0
pbr==6.1.1
pecan==1.5.1
ply==3.11
prettytable==3.14.0
prometheus_client==0.21.1
psutil==6.1.1
pycadf==4.0.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-keystoneclient==5.5.0
pytz==2025.1
PyYAML==6.0.2
radarr-py @ git+https://github.com/devopsarr/radarr-py.git@2c542e8043537cb30f4474c49841a73bea0b1af8
referencing==0.36.2
repoze.lru==0.7
requests==2.32.3
rfc3986==2.0.0
Routes==2.5.1
rpds-py==0.22.3
setuptools==75.8.0
simplegeneric==0.8.1
six==1.17.0
sniffio==1.3.1
sonarr-py @ git+https://github.com/devopsarr/sonarr-py.git@8241df22127af0972163895189478b86a0bffcce
SQLAlchemy==2.0.38
statsd==4.0.1
stevedore==5.4.0
tenacity==9.0.0
testresources==2.0.1
testscenarios==0.5.0
testtools==2.7.2
tooz==6.3.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
WebOb==1.8.9
wrapt==1.17.2
WSME==0.12.1
yappi==1.6.10
yaql==3.0.0
zipp==3.21.0
